refactor pypy-build.sh and add checksum & commit id files

+++ pypy-build.sh
@@ -5,21 +5,39 @@ source pypy-versions.txt
 # PYPY_MAJOR_VERSION
 # PYPY_BUILD_DEP_VERSION
 
+PACKAGE="tklbam-pypy2"
+
 PYPY_REPO="https://github.com/pypy/pypy"
 PYPY_BRANCH="release-pypy2.7-v${PYPY_MAJOR_VERSION}.x"
 
 PYPY_URL="https://downloads.python.org/pypy"
 PYPY_RELEASE_URL="${PYPY_URL}/pypy2.7-v${PYPY_BUILD_DEP_VERSION}-linux64.tar.bz2"
 
-mkdir "pypy-build"
-cd "pypy-build"
-
-BUILD_ROOT="$(pwd)"
+BUILD_ROOT="$(pwd)/build"
 PYPY_SRC="${BUILD_ROOT}/pypy-src"
 PYPY_BIN="${BUILD_ROOT}/pypy-bin"
+PYPY_BUILD="${BUILD_ROOT}/pypy-build"
+
+mkdir "$BUILD_ROOT"
+mkdir "$PYPY_BUILD"
+cd "$BUILD_ROOT"
 
 git clone --depth=1 "${PYPY_REPO}" -b "${PYPY_BRANCH}" pypy-src
+read -r local_commit_id < src-commit-id.txt
+cloned_commit_id=$(git --git-dir=pypy-src/.git rev-parse HEAD)
+if [[ "$local_commit_id" != "$cloned_commit_id" ]]; then
+    echo "source commit ID does not match" >&2
+    exit 1
+fi
+
 curl "$PYPY_RELEASE_URL" -o pypy-bin.tar.bz2
+read -ra local_checksum < checksum.txt
+read -ra dl_checksum <<<"$(sha256sum pypy-bin.tar.bz2)"
+if [[ "${local_checksum[*]}" != "${dl_checksum[*]}" ]]; then
+    echo "Checksums for pypy-bin.tar.bz2 do not match" >&2
+    exit 1
+fi
+
 tar -xvf pypy-bin.tar.bz2
 mv pypy2.7-* pypy-bin
 rm pypy-bin.tar.bz2
@@ -33,18 +51,18 @@ PYPY="${PYPY_BIN}/bin/pypy"
 # --gc=incminimark here is required for the cpyext (or whatever it is, the
 #   thing required for linking C against it) to work.
 #
-# this will error on tkinter, but it's fine, it will do everything it needs too
-"${PYPY}" ../../rpython/bin/rpython --gc=incminimark -Osize targetpypystandalone
+# this will error on tkinter, but it's fine, it will do everything it needs to
+"${PYPY}" ../../rpython/bin/rpython \
+    --gc=incminimark \
+    -Osize targetpypystandalone
 
 cd "${PYPY_SRC}/pypy/tool/release"
-rm -rf /tmp/pypy-build
-mkdir /tmp/pypy-build
 
 "${PYPY}" package.py \
     --without-_tkinter \
     --no-keep-debug \
-    --archive-name pypy-turnkeylinux \
-    --builddir /tmp/pypy-build
+    --archive-name "$PACKAGE" \
+    --builddir "$PYPY_BUILD"
 
 find "/tmp/pypy-build" -type f -iname '*.so' -exec strip -s {} \;
 find "/tmp/pypy-build" -type f -iname '*.so' -exec upx-ucl --best {} \;
